Update tests to use proper handler names
In the real world, the interceptor is always going to be passed a
HandlerCallDetails object with the `method` field set to something
valid... previously these tests set this to an empty string, but that's
not what happens in a real case.

So I've updated the tests to do that right, but also made sure that the
span processing allows this to be unset, just in case someone was
expecting this behavior to be present.