Abort test, clarified attributes
Added a test for handling gRPC aborts, which was previously missing.

Updated the `net.*` attributes, so that `net.peer.ip` is always set, and
`net.peer.name` is set only if it's `localhost`, since otherwise we'd
need to do DNS lookups, and the spec doesn't require that.

Added tests to verify the `net.*` shows up properly.