switch setter for injecting headers w/propagator
Occasionally not all of the headers available to botocore are strings.
It does allow this, using the `add_header` method rather than
`__setitem__`.  According to the docs for `botocore.compat.HTTPHeaders`:

```
If a parameter value contains non-ASCII characters it can be specified
as a three-tuple of (charset, language, value), in which case it will be
encoded according to RFC2231 rules.  Otherwise it will be encoded using
the utf-8 charset and a language of ''
```

Whereas using `__setitem__` does not do this conversion.

Fixes #262