Allow instrumentation of newer FastAPI versions
Since the introduction of the `_instruments` runtime checks in #475, the
FastAPI instrumentation has stopped working for versions >= `0.59.0`.
However the current test suite passes even for the latest released
version at the moment (`0.67.0`).

It seems this isn't related to a limitation in the instrumentation code,
but actually because of it being created when `0.58` was the latest version:
https://github.com/open-telemetry/opentelemetry-python/commit/7bec76a22032e498911d06b9e3dab4e33528d995.