Update filters to handle grpc.aio.ClientCallDetails
This is the precursor for using the filter mechanisms with the aio
interceptors.

There's currently a bug in grpc python that means the the
ClientCallDetails.method field is populated with bytes instead of a
string. This code handles both cases so that it's forward-compatible
with a future fixed grpc version.