revert: feat(instrumentation/asgi): add target to metrics
Unfortunately I made a mistake in https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1323 where I assumed that the scope was laready being processed by the framework at the moment of reporting, but of course that's not true as the framework has not even seen the request at that point.

Is for that reason that we are not able to extract any route information in the middleware to report what the target is.

Sorry for the noise, and be happy to help if anyone has any idea how we could fix this.