Add grpc.aio instrumentation to be auto-instrumented
PR #1245 added grpc.aio instrumentation. It included the auto
instrumentor for grpc.aio but did not include the entry-points hook for
it to be run automatically on boot.

Currently, you need to manually include the instrumentor and call it.
This makes that automatic and in line with the plain grpc
instrumentation.