fix(asgi-instrumentation): extract target after running the framework
There was a mistake in the original strategy to extract the feature, as before running the framework the scope does not have all the information necessary to deduct the route.

We move the target extraction after running the framewrork function, but that means that we won't have those attributes for the span or for the active_requests metrics. The target will be available only for the duration metric.

The test was also updated to reflect that the scope changes only after running the framework function.