Add Redis instrumentation db_statement_serializer
Add user config function `db_statement_serializer` to allow users to
sanitize queries set on the spans on the `db.statement` attribute.

This gives users a way to sanitize queries and not store any PII on the
span.

This user function approach was chosen, because I've seen it in the
Node.js OpenTelemetry instrumentation packages as well.