Fix expected URL in aiohttp instrumentation test
The underlying cause of the issue here is the update of the yarl package
from 1.8.2 to 1.9.1. yarl is used as a dependency in the
opentelemetry-instrumentation-aiohttp package but it is not there where
the issue happens, but in aiohttp who also has yarl as a dependency.

This is why the fix does not touch any relevant part of any
opentelemetry-* code, since it is the return value of aiohttp code who
now has a different value for the URL.

Fixes #1770