Feat: Use the lambda runtime as parent context as opt-in
In a previous PR #1657 the lambda behaviour was changed so that it
extracs the context from the headers of the lambda event instead of from
the lambda runtime, using the _X_AMZN_TRACE_ID env var.

This behaviour is undesireble as it is a breaking change to existing
users.

This PR will add the previous behaviour through an opt-in flag so that
users can be gracefully migrated or these two "modes" of operation can
exist without conflict.

Signed-off-by: Raphael Silva <rapphil@gmail.com>