fix: double call to 'span.end()' when 'acknowledge' is called
The `trace.use_span(span, True)` will already close this span,
if we store this same span in the message, it'll be closed
by a call to any of the `*acknowledge` functions, essentially
causing a double call to `span.end()`.

This change prevents this from happening by not storing the
span in the message.