Fix emitting logs when Logger propagation is disabled for a logger
The logs are emited via `LoggingHandler` as a handler of the root logger.
Unfortunatelly when there is a Logger instance with disabled propagation
with `logger.propagate=False`, the root logger handler does not get called.
Fix this by instrumenting `Logger.propagate` as a property attribute, so
when the value changes to `False`, install the `LoggingHandler` to the
logger automatically, and when the value changes to `True`, uninstall it.

Signed-off-by: Oldřich Jedlička <oldium.pro@gmail.com>