Prevent RuntimeError on context.detach with the same token
This happens for me when using werkzeug's debug=True.
In this case, the token was already detached. But when I enter an
expression into the debug console, we attempt to detach the token once
more.

This causes a RuntimeError, as expected: https://peps.python.org/pep-0567/#contextvars-contextvar.

But if we set _ENVIRON_TOKEN to None after we detach, this won't happen.