Fix tracing of async cursors for psycopg (#3324)
* Fix tracing of async cursors for psycopg

This copies the traced_execution of AsyncCursorTracer except
query_method is awaited within the span.

Fixes https://github.com/open-telemetry/opentelemetry-python-contrib/issues/2486

* Adding a regression test for asynccursor bug

Adds a test to check that the async tracer is actually awaiting
cursor.execute.

On my machine, with the bug present, the duration is 14000ns. With the
bug patched the the duration is orders of magnitude larger.

* Make return type of query_method an awaitable

* Move implementation to traced_execution_async

---------

Co-authored-by: Aaron Abbott <aaronabbott@google.com>